https://github.com/GlowstoneMC/Glowstone/pull/599

+++ src/test/java/net/glowstone/net/ProtocolTestUtils.java
@@ -32,7 +32,7 @@ public final class ProtocolTestUtils {
 
     public static List<MetadataMap.Entry> getMetadataEntry() {
         List<MetadataMap.Entry> list = new ArrayList<>();
-        list.add(new MetadataMap.Entry(MetadataIndex.AGE, 1));
+        list.add(new MetadataMap.Entry(MetadataIndex.AGE, (byte) 1));
         return list;
     }
